+++ src/Service/CodeCommit/src/Exception/EncryptionKeyInvalidIdException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\CodeCommit\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+
+/**
+ * The Key Management Service encryption key is not valid.
+ */
+final class EncryptionKeyInvalidIdException extends ClientException
+{
+}