+++ src/Service/RdsDataService/src/Exception/AccessDeniedException.php
@@ -5,7 +5,7 @@ namespace AsyncAws\RdsDataService\Exception;
 use AsyncAws\Core\Exception\Http\ClientException;
 
 /**
- * You do not have sufficient access to perform this action.
+ * You don't have sufficient access to perform this action.
  */
 final class AccessDeniedException extends ClientException
 {