+++ src/Service/RdsDataService/src/Result/ExecuteStatementResponse.php
@@ -36,12 +36,13 @@ class ExecuteStatementResponse extends Result
     private $numberOfRecordsUpdated;
 
     /**
-     * Values for fields generated during a DML request.```
-     * <note> <p>The <code>generatedFields</code> data isn't supported by Aurora PostgreSQL. To get
-     * the values of generated fields, use the <code>RETURNING</code> clause. For more information, see <a
-     * href="https://www.postgresql.org/docs/10/dml-returning.html">Returning Data From Modified Rows</a>
-     * in the PostgreSQL documentation.</p> </note>
-     * ```.
+     * Values for fields generated during a DML request.
+     *
+     * > The `generatedFields` data isn't supported by Aurora PostgreSQL. To get the values of generated fields, use the
+     * > `RETURNING` clause. For more information, see Returning Data From Modified Rows [^1] in the PostgreSQL
+     * > documentation.
+     *
+     * [^1]: https://www.postgresql.org/docs/10/dml-returning.html
      *
      * @var Field[]
      */